new changes

+++ src/main/java/com/example/DepartmentalStoreCrud/service/CustomerService.java
@@ -17,33 +17,40 @@ public class CustomerService {
         return customRepo.findAll();
     }
 
-    public void addCustomerDetails(Customer customer) {
-        String validEmail = customer.getEmailID();
-        String emailRegex = "^(?=.{1,64}@)[A-Za-z0-9_-]+(\\.[A-Za-z0-9_-]+)*@"
-                + "[A-Za-z0-9]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$";
-
-        if(validEmail.matches(emailRegex)){
-            customRepo.save(customer);
-        }
-        else{
-            throw new IllegalArgumentException("Invalid email");
-        }
-
+    public Customer getCustomerById(Long customerID) {
+        return customRepo.findById(customerID).orElseThrow(NoSuchElementException::new);
     }
 
-    public void updateCustomerDetails(Customer customer) {
-        String validEmail = customer.getEmailID();
+    private void validateEmail(String email) {
         String emailRegex = "^(?=.{1,64}@)[A-Za-z0-9_-]+(\\.[A-Za-z0-9_-]+)*@"
                 + "[A-Za-z0-9]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$";
-        if(validEmail.matches(emailRegex)){
-            customRepo.save(customer);
-        }
-        else{
+
+        if (!email.matches(emailRegex)) {
             throw new IllegalArgumentException("Invalid email");
         }
+    }
+
+    public void addCustomerDetails(Customer customer) {
+        validateEmail(customer.getEmailID());
+        customRepo.save(customer);
+    }
+
+    public void updateCustomerDetails(Long customerID, Customer customer) {
+        Customer existingCustomer = getCustomerById(customerID);
+        if(existingCustomer != null) {
+            existingCustomer.setCustomerID(customerID);
+            existingCustomer.setFullName(customer.getFullName());
+            existingCustomer.setAddress(customer.getAddress());
+            existingCustomer.setContactNumber(customer.getContactNumber());
+            existingCustomer.setEmailID(customer.getEmailID());
+        }
+        validateEmail(existingCustomer.getEmailID());
+        customRepo.save(existingCustomer);
     }
 
     public void deleteCustomerDetails(Long customerID) {
+        Customer customer = getCustomerById(customerID);
+        if(customer != null)
         customRepo.deleteById(customerID);
     }
 }
